+++ lib/providers/course_provider.g.dart
@@ -1,0 +1,27 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'course_provider.dart';
+
+// **************************************************************************
+// RiverpodGenerator
+// **************************************************************************
+
+String _$courseStateNotifierHash() =>
+    r'e056c08d5f75c384eaf3112815750e839af51ca6';
+
+/// See also [CourseStateNotifier].
+@ProviderFor(CourseStateNotifier)
+final courseStateNotifierProvider =
+    AutoDisposeNotifierProvider<CourseStateNotifier, CourseState>.internal(
+  CourseStateNotifier.new,
+  name: r'courseStateNotifierProvider',
+  debugGetCreateSourceHash: const bool.fromEnvironment('dart.vm.product')
+      ? null
+      : _$courseStateNotifierHash,
+  dependencies: null,
+  allTransitiveDependencies: null,
+);
+
+typedef _$CourseStateNotifier = AutoDisposeNotifier<CourseState>;
+// ignore_for_file: type=lint
+// ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member, deprecated_member_use_from_same_package