+++ lib/providers/providers.dart
@@ -4,5 +4,7 @@ export 'srs_lesson_provider.dart';
 export 'feedback_provider.dart';
 export 'user_provider.dart';
 export 'topic_provider.dart';
+export 'course_provider.dart';
+